+++ crucible-llvm/crucible-llvm.cabal
@@ -58,8 +58,9 @@ library
     Lang.Crucible.LLVM.UndefinedBehavior
 
   other-modules:
+    Lang.Crucible.LLVM.Intrinsics.Common
     Lang.Crucible.LLVM.Intrinsics.Libc
-    Lang.Crucible.LLVM.Intrinsics.Types
+    Lang.Crucible.LLVM.Intrinsics.Libcxx
     Lang.Crucible.LLVM.MemModel.Common
     Lang.Crucible.LLVM.MemModel.Generic
     Lang.Crucible.LLVM.MemModel.Value
